[BE] new : add handshake to GPT Server
GPT 서버에 핸드셰이크 과정 추가
- 정상 작동 확인함

+++ ChatGPT Server/app.js
@@ -7,7 +7,7 @@ require("dotenv").config();
 const wss = new WebSocket.Server({ port: process.env.PORT });
 
 // IP를 key로, 웹소켓 배열을 value로 갖는 Map
-let clients = new Map();
+let clients = [];
 
 wss.on("connection", (ws, req) => {
   // 부모님이 설정한 대화를 나누기 위한 프리셋 부분
@@ -29,14 +29,14 @@ wss.on("connection", (ws, req) => {
 
   // 1. 연결 클라이언트 IP 취득
   const ip = req.headers["x-forwarded"] || req.socket.remoteAddress;
-  // 만약 이미 해당 IP의 클라이언트가 있다면
-  if (clients.has(ip)) {
-    // 해당 IP의 웹소켓 배열에 새로운 웹소켓을 추가
-    clients.get(ip).push(ws);
-  } else {
-    // 처음 보는 IP라면 새 배열을 만들고 그 배열에 웹소켓을 추가
-    clients.set(ip, [ws]);
-  }
+  // // 만약 이미 해당 IP의 클라이언트가 있다면
+  // if (clients.has(ip)) {
+  //   // 해당 IP의 웹소켓 배열에 새로운 웹소켓을 추가
+  //   clients.get(ip).push(ws);
+  // } else {
+  //   // 처음 보는 IP라면 새 배열을 만들고 그 배열에 웹소켓을 추가
+  //   clients.set(ip, [ws]);
+  // }
   console.log(`new client[${ip}] connected`);
 
   // 2. 클라이언트에게 메세지 전송(현재는 생략됨)
@@ -48,19 +48,21 @@ wss.on("connection", (ws, req) => {
   ws.on("message", (msg) => {
     let msgJson = JSON.parse(msg);
     console.log(
-      `message from client[${ip}]: ${msgJson.message}, ${msgJson.number}`
+      `message from client[${ip}]:${msgJson.role}, ${msgJson.message}, ${msgJson.number}`
     );
-    // 히스토리에 유저의 답변 저장
-    history.push({
-      role: "user",
-      content: msgJson.message,
-    })(
+
+    if (msgJson.role == "gpt") {
+      // 히스토리에 유저의 답변 저장
+      history.push({
+        role: "user",
+        content: msgJson.message,
+      });
       // 1. gpt에게 답변을 받는다.
       // 2. count를 1증가 시킨다.
       // 3. gpt의 답변 맨 마지막에 ?가 들어 있으면 count를 다시 1 감소 시킨다.
       // 4. count가 randomPoint와 같아졌다면, db를 통해 무작위
-      async () => {
-        let result = await gpt(msgJson.message);
+      (async () => {
+        let result = await gpt(history);
         // 답변속 이모지 제거
         result = deleteEmoji(result);
         // 답변 저장
@@ -71,25 +73,34 @@ wss.on("connection", (ws, req) => {
 
         count = count + 1;
         if (count >= randomPoint) {
-          if (result.charAt(result.length - 1) == "?"){
+          if (result.charAt(result.length - 1) == "?") {
             count = count - 1;
           } else {
-
           }
         }
 
         console.log(`gpt answer: ${result}`);
         // ws.send(`gpt answer: ${result}`);
         // IP가 '특정 IP'인 클라이언트들에게 메시지를 전송
-        if (clients.has(ip)) {
-          clients.get(ip).forEach((client) => {
-            if (client.readyState === WebSocket.OPEN) {
-              client.send(`gpt answer: ${result}`);
-            }
-          });
-        }
-      }
-    )();
+        // if (clients.has(ip)) {
+        //   clients.get(ip).forEach((client) => {
+        //     if (client.readyState === WebSocket.OPEN) {
+        //       client.send(`gpt answer: ${result}`);
+        //     }
+        //   });
+        // }
+        clients.forEach((client) => {
+          if (client.number === ws.number && client.readyState === WebSocket.OPEN) {
+            client.send(`gpt answer: ${result}`);
+          }
+        });
+      })();
+    } else if (msgJson.role == "handshake") {
+      ws.number = msgJson.number;
+      clients.push(ws);
+    } else {
+      ws.send("error: role needed");
+    }
   });
 
   // 4. 에러 처리
@@ -100,15 +111,7 @@ wss.on("connection", (ws, req) => {
   // 5. 연결 종료 이벤트 처리
   ws.on("close", () => {
     // 클라이언트 연결이 닫히면 해당 웹소켓을 배열에서 제거
-    clients.set(
-      ip,
-      clients.get(ip).filter((socket) => socket !== ws)
-    );
-    // 배열에 웹소켓이 하나도 없다면
-    if (clients.get(ip).length === 0) {
-      // 해당 IP를 Map에서 제거
-      clients.delete(ip);
-    }
+    clients = clients.filter((client) => client !== ws);
     console.log(`client[${ip}] connection closed`);
   });
